Rename file chart-section.vue to ChartSection.vue
