+++ components/ChartSection.vue
